Update listview.md
Altered the opening line of the remarks, because it confusingly characterizes ListView.  The ListView's ItemsPanel determines whether the items are vertical or horizontal, and is easily customizable; the _default_ template uses a vertical ItemsStackPanel, but folks should still be encouraged to use ListView if they want the collection arranged horizontally.